+++ test/array.js
@@ -1,56 +1,124 @@
 module('Array');
   
 test('Array.intersect', function() {
+  strictEqual(Array.intersect.name, 'intersect');
   deepEqual(Array.intersect([[1,2,3],[2,3,4],[3,4,5]]), [3]);
+  deepEqual(Array.intersect([1,2,3],[2,3,4],[3,4,5]), [3]);
+  deepEqual(Array.intersect([]), []);
+  deepEqual(Array.intersect(), []);
 });
   
 test('Array.diff', function() {
+  strictEqual(Array.diff.name, 'diff');
   deepEqual(Array.diff([[1,2,3],[2,3,4],[3,4,5]]), [1,5]);
+  deepEqual(Array.diff([1,2,3],[2,3,4],[3,4,5]), [1,5]);
+  deepEqual(Array.diff([]), []);
+  deepEqual(Array.diff(), []);
 });
   
 test('Array.union', function() {
+  strictEqual(Array.union.name, 'union', 'Function Name');
   deepEqual(Array.union([[1,2,3],[2,3,4],[3,4,5]]), [1,2,3,4,5]);
+  deepEqual(Array.union([1,2,3],[2,3,4],[3,4,5]), [1,2,3,4,5]);
+  deepEqual(Array.union([]), []);
+  deepEqual(Array.union(), []);
 });
   
 test('Array.range', function() {
   var $R = Array.range;
-  deepEqual($R(3), [1,2,3]);
-  deepEqual($R(1,5,2), [1,3,5]);
+  strictEqual($R.name, 'range');                   // 1
+  deepEqual($R(3), [1,2,3]);                       // 2
+  deepEqual($R('a','a'), []);                      // 3
+  deepEqual($R(1,5,2), [1,3,5]);                   // 4
+  deepEqual($R(1,5), [1,2,3,4,5]);                 // 5
+  deepEqual($R('3',5,'1'), [3,4,5]);               // 6
+  deepEqual($R(3,-2,-2), [3,1,-1]);                // 7
+  deepEqual($R(-3), [-1,-2,-3]);                   // 8
+  raises(function() { $R(1,2,0) }, RangeError);    // 9
+  raises(function() { $R(1,2,-1) }, RangeError);   // 10
+  raises(function() { $R(1,-1,1) }, RangeError);   // 11
+  raises(function() { $R(-1,1,'a') }, RangeError); // 12
 });
 
 test('Array#concat$', function() {
   var arr = [1,2,3];
-  strictEqual(arr.concat$([4,5,6]), arr);
+  strictEqual(arr.concat$.name, 'concat$');
+  strictEqual(arr.concat$(4,5,6), arr);
   deepEqual(arr, [1,2,3,4,5,6]);
+  deepEqual(arr.concat$(), [1,2,3,4,5,6]);
+});
+
+test('Array#swap', function() {
+  // Rest of the tests are covered by swap$
+  var arr = [1,2,3];
+  strictEqual(arr.swap.name, 'swap');
+  ok(arr.swap(2, 0) !== arr);
+  deepEqual(arr.swap(2, 0), [3,2,1]);
 });
   
-test('Array#swap', function() {
-  deepEqual([1,2,3].swap(2, 0), [3,2,1]);
+test('Array#swap$', function() {
+  var arr = [1,2,3];
+  strictEqual(arr.swap$.name, 'swap$');                         // 1
+  strictEqual(arr.swap$(2,0), arr);                             // 2
+  deepEqual(arr,[3,2,1]);                                       // 3
+  deepEqual([1,2,3].swap$(2,'0'), [3,2,1]);                     // 4
+  raises(function() { [1,2,3].swap$(1,3) }, RangeError);        // 5
+  raises(function() { [1,2,3].swap$(-1,2) }, RangeError);       // 6
+  raises(function() { [1,2,3].swap$() }, RangeError);           // 7
+  raises(function() { [1,2,3].swap$(1,'length') }, RangeError); // 8
+  
+  arr.foo = 'bar';
+  raises(function() { arr.swap$(0,'foo') }, RangeError);        // 9
+  
+  arr = new Array(5);
+  arr[3] = 2, arr[1] = 1;
+  deepEqual(arr.swap$(1,3), [,2,,1,,]);                         // 10
 });
   
 test('Array#contains', function() {
-  ok([1,2,3].contains(1,2));
-  equal([1,2,3].contains(4), false);
+  strictEqual([].contains.name, 'contains');    // 1
+  ok([1,2,3].contains(1,2));                    // 2
+  strictEqual([1,2,3].contains(4), false);      // 3
+  strictEqual([1,2,3].contains(), false);       // 4
+  strictEqual([].contains(2), false);           // 5
 });
-  
+
 test('Array#remove', function() {
   var a = [1,2,3];
-  equal(a.remove(2,3), a);
-  deepEqual(a, [1]);
+  strictEqual(a.remove.name, 'remove');
+  ok(a.remove(2,3) !== a);
+  deepEqual(a.remove(2,3), [1,2,3].remove$(2,3));
+  // Further testing is done in #remove$
+});
+  
+test('Array#remove$', function() {
+  
+  var arr = [1,2,3];
+  arr.foo = 'bar';
+
+  strictEqual(arr.remove$.name, 'remove$');
+  deepEqual(arr.remove$(2,3), [1]);
+  
+  arr.foo = 'bar';
+  strictEqual(arr.remove$('foo').foo, 'bar');
+  
+  deepEqual([].remove$(1,2), []);
 });
   
 test('Array#shuffle', function() {
   var a = [1,2,3], b = a.shuffle();
+  strictEqual(a.shuffle.name, 'shuffle');
   notStrictEqual(b, a);
   ok(a.contains(1,2,3));
-  //TODO: Test shuffling mechanism
 });
   
 test('Array#shuffle$', function() {
   var a = [1,2,3], b = a.shuffle$();
+  strictEqual(a.shuffle$.name, 'shuffle$');
   strictEqual(a, b);
-  //TODO: Test shuffling mechanism
+  ok(a.contains(1,2,3));
 });
+
   
 test('Array#clone', function() {
   var a = [1,2,3], b = a.clone();
@@ -139,6 +207,7 @@ test('Array#filter$', function() {
   var a = [1,2,3,4,5,6];
   strictEqual(a.filter$(function(a) { return a.even() }), a);
   deepEqual(a, [2,4,6]);
+  deepEqual([,,,].filter(function(){return true}), []);
 });
   
 test('Array#map$', function() {